Do not pass whole sequence to callbacks
