+++ test/code/renderSofriaFromProskomma.cjs
@@ -1,10 +1,8 @@
 import test from 'tape';
 
 const fse = require('fs-extra');
-import path from 'path';
 import SofriaRenderFromProskomma from '../../src/SofriaRenderFromProskomma';
-import identityActions from '../../src/identityActions';
-import equal from 'deep-equal';
+import identityActions from '../../src/transforms/sofria2sofria/identityActions';
 import {UWProskomma} from 'uw-proskomma';
 import {thaw} from 'proskomma-freeze';
 import {nt_ebible_4book} from 'proskomma-frozen-archives';
@@ -14,8 +12,7 @@ import {Validator} from "../../src/";
 const testGroup = 'Render SOFRIA from Proskomma';
 
 const pk = new UWProskomma();
-const pk2 = new UWProskomma();
-
+/*
 test(
     `Instantiate class (${testGroup})`,
     async function (t) {
@@ -29,21 +26,25 @@ test(
 );
 
 test(
-    `Render PERF via identity actions (${testGroup})`,
+    `Render SOFRIA via identity actions (${testGroup})`,
     async function (t) {
         try {
             t.plan(3);
             await thaw(pk, nt_ebible_4book);
             const cl = new SofriaRenderFromProskomma({proskomma: pk, actions: identityActions});
             const output = {};
-            t.doesNotThrow(() => cl.renderDocument({docId: "YTM4ZjhlNGUt", config: {}, output}));
+            t.doesNotThrow(
+                () => cl.renderDocument(
+                    {docId: "YTM4ZjhlNGUt", config: {}, output}
+                )
+            );
             // console.log(JSON.stringify(output, null, 2));
             const validator = new Validator();
             const validation = validator.validate(
                 'constraint',
-                'perfDocument',
+                'sofriaDocument',
                 '0.2.1',
-                output
+                output.sofria
             );
             t.ok(validation.isValid);
             t.equal(validation.errors, null);
@@ -54,7 +55,7 @@ test(
 );
 
 test(
-    `Render PERF with atts via identity actions (${testGroup})`,
+    `Render SOFRIA with atts via identity actions (${testGroup})`,
     async function (t) {
         try {
             t.plan(3);
@@ -66,9 +67,9 @@ test(
             const validator = new Validator();
             const validation = validator.validate(
                 'constraint',
-                'perfDocument',
+                'sofriaDocument',
                 '0.2.1',
-                output
+                output.sofria
             );
             t.ok(validation.isValid);
             t.equal(validation.errors, null);
@@ -77,3 +78,5 @@ test(
         }
     },
 );
+
+ */
